wip: node physics

+++ crates/common/src/geometry.rs
@@ -1,8 +1,8 @@
 /// Line between two points
-pub type Line = (mint::Point2<f64>, mint::Point2<f64>);
+pub type Line = (mint::Point2<i64>, mint::Point2<i64>);
 
 /// Rectangle from start of coordinates
 ///
-/// - start `mint::Point2<f64>`
-/// - size `mint::Vector2<f64>`
-pub type Rect = (mint::Point2<f64>, mint::Vector2<f64>);
+/// - start `mint::Point2<i64>`
+/// - size `mint::Vector2<i64>`
+pub type Rect = (mint::Point2<i64>, mint::Vector2<i64>);
